Check for user ownership for todos and tasks

+++ walking-skeleton/server/taskRepository.js
@@ -2,7 +2,12 @@ import postgres from "postgres";
 
 const sql = postgres();
 
-const createTask = async (todoId, task) => {
+
+const createTask = async (userId, todoId, task) => {
+  const getTodo = await sql`SELECT * FROM todos WHERE todos.id = ${todoId} AND todos.user_id = ${userId}`;
+  if (getTodo.length === 0){
+    return null;
+  };
   const result = await sql`
     INSERT INTO todo_tasks (todo_id, description)
     VALUES (${todoId}, ${task.description})
@@ -11,7 +16,11 @@ const createTask = async (todoId, task) => {
   return result[0];
 };
 
-const deleteById = async ( id) => {
+const deleteById = async (userId, id) => {
+   const userTodoTaskCount = await sql`SELECT COUNT(*) as count FROM todo_tasks JOIN todos ON todo_tasks.todo_id = todos.id WHERE todo_tasks.id = ${id} AND todos.user_id = ${userId}`;
+  if (userTodoTaskCount[0]?.count === "0"){
+    return null;
+  };
   const result = await sql`
     DELETE FROM todo_tasks
     WHERE  id = ${id}
@@ -20,13 +29,21 @@ const deleteById = async ( id) => {
   return result[0];
 };
 
-const findTasks = async (todoId) => {
+const findTasks = async (userId, todoId) => {
+   const getTodo = await sql`SELECT * FROM todos WHERE todos.id = ${todoId} AND todos.user_id = ${userId}`;
+  if (getTodo.length === 0){
+    return null;
+  };
   return await sql`
     SELECT * FROM todo_tasks
     WHERE todo_id = ${todoId};`;
 };
 
-const findById = async (id) => {
+const findById = async (userId, id) => {
+   const userTodoTaskCount = await sql`SELECT COUNT(*) as count FROM todo_tasks JOIN todos ON todo_tasks.todo_id = todos.id WHERE todo_tasks.id = ${id} AND todos.user_id = ${userId}`;
+  if (userTodoTaskCount[0]?.count === "0"){
+    return null;
+  };
   const result = await sql`
     SELECT * FROM todo_tasks
     WHERE id = ${id};`;
@@ -34,7 +51,11 @@ const findById = async (id) => {
   return result[0];
 };
 
-const updateById = async (id, task) => {
+const updateById = async (userId, id, task) => {
+  const userTodoTaskCount = await sql`SELECT COUNT(*) as count FROM todo_tasks JOIN todos ON todo_tasks.todo_id = todos.id WHERE todo_tasks.id = ${id} AND todos.user_id = ${userId}`;
+  if (userTodoTaskCount[0]?.count === "0"){
+    return null;
+  };
   const result = await sql`
     UPDATE todo_tasks
     SET
